Creating Worker component class and IWorkstation
IWorkstation can be used by an IWorker. Also created the CraftingPad in the scene
with defaults set up for crafting a drone. Workstations can accept any number of IWorkers,
with a max set in the inspector. The Worker ability has been added to the drone.

The repacked.dll has been removed as this is referred to in the install directory directly by unity 2019.2 onwards (this is a Rider plugin).

DataManager added (with ResourceData) but is not currently being used